requirements.txt
python-telegram-bot>=20.0
SQLAlchemy>=2.0
asyncpg
alembic
python-dotenv
psycopg2-binary
saq[redis]
redis 